站点，校车信息，学生等在delete时不删除，只把valid设为0

+++ src/main/java/com/eservice/api/model/bus_base_info/BusBaseInfo.java
@@ -55,7 +55,7 @@ public class BusBaseInfo {
     @Column(name = "ipad_meid")
     private String ipadMeid;
 
-    private Byte valid;
+    private Integer valid;
 
     /**
      * @return id
@@ -218,14 +218,14 @@ public class BusBaseInfo {
     /**
      * @return valid
      */
-    public Byte getValid() {
+    public Integer getValid() {
         return valid;
     }
 
     /**
      * @param valid
      */
-    public void setValid(Byte valid) {
+    public void setValid(Integer valid) {
         this.valid = valid;
     }
 }
